feat: add cache for storage configuration (#960)
Added cache for storage. At first, it reads storage section from Artipie settings. If this content exists in cache and is not expired, storage will be loaded from the cache. Otherwise a new instance will be created.
In next PR I'm planning to use this cache in YamlSettings#storage

Ticket: #87