refactor: extract cache package and logic of reading property value (#963)
Extracted cache package.
Extracted logic of reading value of the property from environment and Artipie properties file.
Fixed reading value of property and added tests for it. Now it takes value of property from environment, if it was set, otherwise it looks up in artipie.properties file. If both of them are null it parses to long passed default value.

Ticket: #87